requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.55.3
greenlet==3.1.1
importlib-resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
typing-extensions==4.12.2
tzdata==2024.2
zipp==3.20.2